Refactor and optimize
Refactor and optimize the codebase

* **Configuration Handling:**
  - Add `src/disopy/config_loader.py` to handle loading the configuration.
  - Add `src/disopy/config_saver.py` to handle saving the configuration.
  - Move `get_config`, `get_config_file_path`, and `generate_new_config_tip_message` functions to `src/disopy/config_loader.py`.
  - Move `generate_new_config` function to `src/disopy/config_saver.py`.
  - Remove the `update_volume` function from `src/disopy/config.py`.
  - Import necessary functions from `src/disopy/config_loader.py` and `src/disopy/config_saver.py` in `src/disopy/config.py`.

* **Queue Management:**
  - Refactor `Queue` class in `src/disopy/cogs/queue.py` to separate queue management and guild checking logic.
  - Create a new `GuildChecker` helper class for guild checking logic.
  - Update `Queue` class to use `GuildChecker` for guild-related operations.

**Double check my code cause I might've misconfigured my testing environment although it did run.**
