+++ src/disopy/config.py
@@ -6,12 +6,10 @@
 
 import logging
 from pathlib import Path
-from typing import Any, NamedTuple
+from typing import NamedTuple
 
-import tomlkit
-from tomlkit import comment, document, item, nl, table
-
-from . import APP_NAME, APP_NAME_LOWER
+from .config_loader import get_config, get_config_file_path
+from .config_saver import generate_new_config
 
 logger = logging.getLogger(__name__)
 
@@ -38,14 +36,6 @@ class Config(NamedTuple):
 
     developer_discord_sync_guild: int | None
     developer_discord_sync_users: list[int]
-
-    def update_volume(self, new_volume: int) -> None:
-        """Update the volume setting in the config.
-
-        Args:
-            new_volume: The new volume level.
-        """
-        self.volume = new_volume
 
     def save(self, config_path: Path) -> None:
         """Save the updated config to the config file.
@@ -77,131 +67,3 @@ class Config(NamedTuple):
 
         with open(config_file, "w") as f:
             f.write(doc.as_string())
-
-
-def get_config_file_path(config_path: Path) -> Path:
-    """Get the config file path and generate the file.
-
-    Args:
-        config_path: The path of the config folder.
-
-    Returns:
-        The path of the config file.
-    """
-
-    config_path.mkdir(parents=True, exist_ok=True)
-
-    config_file = config_path / "config.toml"
-    config_file.touch()
-
-    return config_file
-
-
-def generate_new_config_tip_message() -> None:
-    """Print a tip about how to generate a default config"""
-
-    logger.info(f"You can generate a default config file with: {APP_NAME_LOWER} --generate-config")
-
-
-def missing_entry_error_message(missing_entry: str) -> None:
-    """Print an error about a missing entry in the config file"""
-
-    logger.critical(f"The provided config file is missing the '{missing_entry}' entry")
-    generate_new_config_tip_message()
-
-
-def generate_new_config(config_path: Path) -> None:
-    """Generate a new config file
-
-    Args:
-        config_path: The path to the folder where config file should be stored.
-    """
-
-    logger.info("Generating a new config file")
-
-    config_file = get_config_file_path(config_path)
-
-    doc = document()
-    doc.add(comment(f"{APP_NAME} config file"))
-    doc.add(nl())
-
-    doc.add(comment("DO NOT MODIFY ME! Internal config file version"))
-    # Ugly hack to avoid `mypy` errors when using numbers, the `tomlkit` typing is really bad
-    doc.add("version", item(1))
-    doc.add(comment("The volume of the music playback in percentage"))
-    doc.add("volume", item(100))
-
-    doc.add(nl())
-
-    subsonic_table = table()
-    subsonic_table.add(comment("The URL where the OpenSubsonic REST API can be accessed"))
-    subsonic_table.add("url", "http://example.com")
-
-    subsonic_table.add(comment("The user to be used when authenticating in the OpenSubsonic server"))
-    subsonic_table.add("user", "Alice")
-
-    doc.add("subsonic", subsonic_table)
-
-    with open(config_file, "w") as f:
-        f.write(doc.as_string())
-
-
-def get_config(config_path: Path) -> Config | None:
-    """Process the config files and returns it deserialized.
-
-    Args:
-        config_path: The path to the folder where config file should be accessed.
-    """
-
-    config_file = get_config_file_path(config_path)
-
-    with open(config_file, "r+") as f:
-        try:
-            config: dict[str, Any] = tomlkit.load(f)
-        except tomlkit.exceptions.ParseError as e:
-            logger.critical("The provided config file is not valid TOML")
-            generate_new_config_tip_message()
-            logger.critical(e)
-
-        if "version" not in config:
-            missing_entry_error_message("version")
-            return None
-
-        version = int(config["version"])
-
-        # Here old version can be migrated
-
-        if "volume" not in config:
-            missing_entry_error_message("volume")
-            return None
-
-        volume = int(config["volume"])
-        if volume < 0:
-            logger.critical("The volume config entry is lower than 0%")
-
-        if "subsonic" not in config:
-            missing_entry_error_message("subsonic")
-            return None
-
-        for entry in ["url", "user"]:
-            if entry not in config["subsonic"].value:
-                missing_entry_error_message(f"subsonic.{entry}")
-                return None
-
-        subsonic_url = str(config["subsonic"]["url"])
-        subsonic_user = str(config["subsonic"]["user"])
-
-        developer_discord_sync_guild = None
-        developer_discord_sync_users: list[int] = []
-        if "developer" in config:
-            developer_discord_sync_guild = config["developer"]["discord-sync-guild"]
-            developer_discord_sync_users = config["developer"]["discord-sync-users"]
-
-        return Config(
-            version=version,
-            volume=volume,
-            subsonic_url=subsonic_url,
-            subsonic_user=subsonic_user,
-            developer_discord_sync_guild=developer_discord_sync_guild,
-            developer_discord_sync_users=developer_discord_sync_users,
-        )
